requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
PyGithub
python-dotenv
llama-index-core
llama-index-llms-openai
